python/requirements.txt
Django>=5.0
PyYAML>=6.0
Markdown>=3.5
python-dotenv>=1.0